Removed faulty call to Dispose from HardwareBufferManager( cause of a StackOverflowException ); First branch of changes concerning undisposed objects;

+++ Source/Core/Axiom/CrossPlatform/Buffer.cs
@@ -213,7 +213,7 @@ namespace Axiom.Core
 			}
 
 			var buf = (BufferBase)buffer.Clone();
-			buf.Ptr += (int)offset;
+			buf.Ptr += offset;
 			return buf;
 		}
